Convert Wrapper component to typescript
